Double speed in release mode through forced inlining
Speedup:
- **Unrolled loops in permutations**, so the ascon_round is
  called directly 6/8/12 times, each with a different
  macro for a round constant
- Removed static array of round constants, because they are now
  macros
This results in 2x the speed on long data (32kB), being now about
2x slower than the reference implementation instead of 4x slower.

Additionally the round function and permutation functions are
now inlined optionally: the ASCON_INLINE macro becomes
"inline" only when the release build is selected. In debug
mode and minsizerel mode the inline is not suggested to the
compiler. For this reason minsizerel is smaller than the
release but 2x slower as release (thus exactly the same
speed as before)

Debug:     292400 B
Minsizerel: 78584 B
Release:    88584 B

Reference
|     1 |     8 |    16 |    32 |    64 |  1536 | 32768 |
|------:|------:|------:|------:|------:|------:|------:|
|   463 |    75 |    45 |    29 |  21.1 |  13.9 |  13.4 |

Release
|     1 |     8 |    16 |    32 |    64 |  1536 | 32768 |
|------:|------:|------:|------:|------:|------:|------:|
|   459 |    80 |    51 |    37 |  30.0 |  23.5 |  23.0 |

MinSizeRel
|     1 |     8 |    16 |    32 |    64 |  1536 | 32768 |
|------:|------:|------:|------:|------:|------:|------:|
|   947 |   148 |    94 |    67 |  53.6 |  40.8 |  40.0 |