Reach same performance as original implementation (!!)
Basically LibAscon was already as performant as the original
reference implementation, but the benchmarking test was flawed.
Now the benchmarking in release mode produces identical results.

For yet unknown reasons the rewritten benchmarking produces
about 2x slower results. By replacing the whole benchmark.c
with the content of the original and just calling the LibAscon
function instead of the original one, the performances are
exactly the same.