Suppress not relevant warnings about inlining
They appear for the testascon build where the Ascon library
is linked to the test executable statically, which makes the
compiler inline also some of the Ascon API functions, although
they were not marked as inline-able. This ia purely informational
warning and anyhow appears only for the test suite code, thus
removed.