Fix GCC v11 warnings about uint8* vs uint8[]
Difference in function signature in declaration and implementation.
Apparently now the compiler wants them to be consistent, so they now
are. Arrays are used as data type (although they are just
pointers behind the scenes), so the known, constant length of the
binary key/nonce can be explicitly indicated for code clarity,
following the recommended practice from the MISRA-C standard.